Handle case where objects not present in all rounds.
Consensus table has NaN values for objects that are present in reference round but not in certain RX rounds.
Give option for user to either discard these objects or to preserve them (and in this case ignore NaNs).
If the former, this means that numerical order of relabeled objects might have gaps (e.g. 1,2,3,5, with 4 skipped if the object was
not found in all rounds).